+++ src/configWoosmapGeofencingIOS.ts
@@ -2,6 +2,7 @@ import {
   ConfigPlugin,
   withInfoPlist,
   withDangerousMod,
+  withXcodeProject,
 } from "@expo/config-plugins";
 
 import { ConfigProps } from "./types";
@@ -94,10 +95,50 @@ const withSDKDangerousMod: ConfigPlugin<ConfigProps> = (config, props) => {
     },
   ]);
 };
+
+
+const withSDKXcodeProject: ConfigPlugin<ConfigProps> =  (config,props) => {
+  
+  return withXcodeProject(config, async (config) => {
+
+    const xcodeProject = config.modResults;
+    const PHASE_NAME = 'Remove WoosmapGeofencing signature file';
+    const shellScript = `echo "Remove WoosmapGeofencing signature file"
+  rm -rf "$BUILD_DIR/\${CONFIGURATION}-iphoneos/WoosmapGeofencing.xcframework-ios.signature"`;
+
+    // -----------------------------
+    // 1) Prevent duplicate insertion
+    // -----------------------------
+    const existingPhases = xcodeProject.pbxItemByComment(PHASE_NAME, 'PBXShellScriptBuildPhase');
+    if (existingPhases) {
+      console.log(`⚠️ '${PHASE_NAME}' phase already exists, skipping creation.`);
+      return config;
+    }
+
+
+    const phaseUUID =  xcodeProject.addBuildPhase(
+      [],
+      'PBXShellScriptBuildPhase',
+      PHASE_NAME,
+      null,
+      {
+        shellPath: '/bin/sh',
+        shellScript,
+      },
+    );
+
+    // Patch the field manually
+    const phase = xcodeProject.pbxItemByComment(PHASE_NAME, 'PBXShellScriptBuildPhase');//pbxShellScriptBuildPhaseObj(phaseUUID);
+    phase.runOnlyForDeploymentPostprocessing = 1;
+    
+    return config;
+  });
+};
+
 export const withIOSSdk: ConfigPlugin<ConfigProps> = (config, props) => {
   config = withSDKInfoPlist(config, props);
   //   config = withSDKEntitlements(config, props);
-  //   config = withSDKXcodeProject(config, props);
+  config = withSDKXcodeProject(config,props);
   config = withSDKDangerousMod(config, props);
   return config;
 };